Adjust helper scripts to detect if we need to use the new docker compose command instead of the older docker-compose command.  To use new format, uninstall the docker-compose command (apt remove docker-compose or pip uninstall docker-compose)

+++ docker_build_files/up.sh
@@ -4,16 +4,21 @@
 
 PY3=`which python3`
 PY2=`which python`
+compose=`which docker-compose`
 
 PY=$PY3
 
 if [ -z "$PY" ]; then
-  echo Switching to py2
+  #echo Switching to py2
   PY=$PY2
 fi;
 
 echo "Using Python: $PY"
 
+if [ -z "$compose" ]; then
+  compose="docker compose"
+fi;
+echo "Using Compose: $compose"
 
 SCRIPT=$(readlink -f "$0")
 BASEDIR=$(dirname "$SCRIPT")
@@ -21,37 +26,37 @@ ROOTDIR=$(dirname "$BASEDIR")
 
 cd "$BASEDIR"
 
-# Only run this stuff if fog is enabled
+# Only run this stuff if fog is enabled and running in old opensuse OS
 if [ -f ope-fog/.enabled ]; then
+    if /usr/bin/grep -q "SUSE" "/etc/os-release"; then
+      systemctl disable rpcbind
+      systemctl stop rpcbind
 
-    systemctl disable rpcbind
-    systemctl stop rpcbind
+      # change to force update/checkout after adjusting gitattribute linefeeds
 
-    # change to force update/checkout after adjusting gitattribute linefeeds
-
-    # Need these loaded for tftp server to work
-    echo "Ensuring kernel modules are loaded..."
-    modprobe nf_conntrack_tftp
-    echo "nf_conntrack_tftp" > /etc/modules-load.d/nf_conntrack_tftp.conf
-    modprobe nf_nat_tftp
-    echo "nf_nat_tftp" > /etc/modules-load.d/nf_nat_tftp.conf
-    modprobe nf_conntrack_ftp
-    echo "nf_conntrack_ftp" > /etc/modules-load.d/nf_conntrack_ftp.conf
-    modprobe nf_conntrack_netbios_ns
-    echo "nf_conntrack_netbios_ns" > /etc/modules-load.d/nf_conntrack_netbios_ns.conf
-    modprobe nfs
-    echo "nfs" > /etc/modules-load.d/nfs.conf
-    modprobe nfsd
-    echo "nfsd" > /etc/modules-load.d/nfsd.conf
-    modprobe ipip
-    echo "ipip" > /etc/modules-load.d/ipip.conf
+      # Need these loaded for tftp server to work
+      echo "Ensuring kernel modules are loaded..."
+      modprobe nf_conntrack_tftp
+      echo "nf_conntrack_tftp" > /etc/modules-load.d/nf_conntrack_tftp.conf
+      modprobe nf_nat_tftp
+      echo "nf_nat_tftp" > /etc/modules-load.d/nf_nat_tftp.conf
+      modprobe nf_conntrack_ftp
+      echo "nf_conntrack_ftp" > /etc/modules-load.d/nf_conntrack_ftp.conf
+      modprobe nf_conntrack_netbios_ns
+      echo "nf_conntrack_netbios_ns" > /etc/modules-load.d/nf_conntrack_netbios_ns.conf
+      modprobe nfs
+      echo "nfs" > /etc/modules-load.d/nfs.conf
+      modprobe nfsd
+      echo "nfsd" > /etc/modules-load.d/nfsd.conf
+      modprobe ipip
+      echo "ipip" > /etc/modules-load.d/ipip.conf
 
 
-    # Add some rules to track tftp traffic
-    #WLAN_IF=eth0
-    #iptables -A INPUT -i $WLAN_IF -p udp -m state --state ESTABLISHED,RELATED -j ACCEPT
-    #iptables -A INPUT -i $WLAN_IF -p udp --dport 69 -m state --state NEW -j ACCEPT
-
+      # Add some rules to track tftp traffic
+      #WLAN_IF=eth0
+      #iptables -A INPUT -i $WLAN_IF -p udp -m state --state ESTABLISHED,RELATED -j ACCEPT
+      #iptables -A INPUT -i $WLAN_IF -p udp --dport 69 -m state --state NEW -j ACCEPT
+    fi
 fi
 
 
@@ -64,17 +69,20 @@ rebuild_param=""
 if [ "$build_flag" = "auto" ]; then
   rebuild_param="auto"
 fi
-echo "Rebuilding docker compose..."
+#echo "Rebuilding docker compose..."
 $PY $ROOTDIR/build_tools/rebuild_compose.py "$rebuild_param"
 
 if [ "$build_flag" = "b" ]; then
   echo "Building docker containers..."
-  docker-compose build
+  #docker-compose build
+  $compose build
 fi
 
 echo "Bringing up containers..."
 # Bring up without rebuilding (in case container is out of date it will still start)
-docker-compose up -d --no-build --remove-orphans
+#docker-compose up -d --no-build --remove-orphans
+$compose up -d --no-build --remove-orphans
+
 
 
 #echo "Bringing up bridge for fog..."
